refactoring
extracted addresses factory
